Rename wait module to wait_set
