+++ rclrs/src/lib.rs
@@ -20,7 +20,7 @@ mod subscription;
 mod time;
 mod time_source;
 mod vendor;
-mod wait;
+mod wait_set;
 
 #[cfg(test)]
 mod test_helpers;
@@ -45,4 +45,4 @@ pub use service::*;
 pub use subscription::*;
 pub use time::*;
 use time_source::*;
-pub use wait::*;
+pub use wait_set::*;
